Refactor after reveiw
