+++ src/components/BaseQuestionnaireResponseForm/widgets/GroupTable/GroupTableChart/HighlightArea/index.ts
